Split match_report.py and overlay/routes.py; drop C901 suppressions
match_report.py (2,094 lines) becomes a thin route module backed by
three siblings following the existing convention:
- match_report_access.py — read/delete auth ladder
- match_report_stats.py — pure audit-log reducers (live_stats now
  imports from here, which also breaks a latent import cycle through
  app.api)
- match_report_render.py — HTML/SVG fragment builders

overlay/routes.py sheds require_overlay_server_token (overlay/auth.py,
re-exported for the documented import path), locale resolution
(overlay/locale.py) and the Pydantic models (overlay/models.py); route
registration moves into _register_page_routes/_register_api_routes
helpers.

The four over-complex functions (_compute_stats, _render_highlights,
_render_score_chart, create_overlay_router) are decomposed below the
mccabe cap, so both per-file C901 ignores in pyproject.toml are
deleted. Route paths, operation ids and behaviour unchanged — the
regenerated OpenAPI snapshot is byte-identical.
