+++ app/match_report.py
@@ -33,25 +33,44 @@ unlock public *delete* — the operator has to opt in to each.
 
 from __future__ import annotations
 
-import datetime
 import html
 import logging
-import re
 import time
-from typing import Any, overload
 
 from fastapi import APIRouter, Header, HTTPException, Query, Response
 from fastapi.responses import HTMLResponse
 
 from app.api import match_archive
-from app.api.schemas import ERROR_TYPES, POINT_TYPES
-from app.auth_utils import require_admin_token as _require_admin_token
-from app.env_vars_manager import EnvVarsManager
-from app.match_report_i18n import (
-    SUPPORTED_LOCALES,
-    resolve_locale,
+from app.match_report_access import (
+    _check_access,
+    _check_admin_access,
+    _public_delete_enabled,
 )
+from app.match_report_i18n import SUPPORTED_LOCALES, resolve_locale
 from app.match_report_i18n import t as _t
+from app.match_report_render import (
+    _INDEX_SCRIPT,
+    _INDEX_TEMPLATE,
+    _chart_color,
+    _ensure_distinct_chart_colors,
+    _fmt_seconds,
+    _fmt_ts,
+    _render_charts,
+    _render_highlights,
+    _render_logo,
+    _render_match_row,
+    _render_set_durations_row,
+    _render_timeline,
+    _team_color,
+    _team_name,
+)
+from app.match_report_stats import (
+    _collapse_undos,
+    _compute_stats,
+    _played_set_count,
+    _timeouts_per_set,
+    _trim_pregame,
+)
 from app.match_report_template import REPORT_TEMPLATE as _REPORT_TEMPLATE
 
 
@@ -72,1463 +91,6 @@ def _is_supported_locale_tag(value: str | None) -> bool:
 logger = logging.getLogger(__name__)
 
 match_report_router = APIRouter()
-
-
-def _public_mode_enabled() -> bool:
-    return EnvVarsManager.get_bool_env("MATCH_REPORT_PUBLIC")
-
-
-def _public_delete_enabled() -> bool:
-    """``True`` when the operator has opted into unauthenticated delete.
-
-    Independent from :func:`_public_mode_enabled` on purpose: granting
-    public read shouldn't silently authorise public destruction.
-    """
-    return EnvVarsManager.get_bool_env("MATCH_REPORT_PUBLIC_DELETE")
-
-
-def _check_access(
-    authorization: str | None,
-    token: str | None,
-    *,
-    match_id: str | None = None,
-    exp: str | None = None,
-    sig: str | None = None,
-) -> None:
-    """Raise an ``HTTPException`` unless the caller is allowed to read.
-
-    Order of precedence:
-      1. ``MATCH_REPORT_PUBLIC=true`` — open access (matches the
-         existing ``/overlay/{output_key}`` model);
-      2. a valid HMAC signature on ``(match_id, exp)`` (capability URL
-         minted by the admin endpoint, no password in the link);
-      3. otherwise, ``OVERLAY_MANAGER_PASSWORD`` must be set and
-         provided via Bearer header or ``?token=`` query (legacy);
-      4. when neither password nor signature works and no public-read
-         mode is enabled, return 503 to make misconfiguration loud
-         rather than silently public.
-    """
-    if _public_mode_enabled():
-        return
-    # Capability URL — no need for the password to be on the wire.
-    # Signed URLs are minted by the admin endpoint and embed an
-    # ``exp`` that lets the operator share a short-lived link
-    # without leaking ``OVERLAY_MANAGER_PASSWORD``.
-    if match_id is not None and (exp or sig):
-        from app.match_report_signing import verify_signed_query
-        if verify_signed_query(match_id, exp, sig):
-            return
-        # Falling through is intentional: an invalid sig should not
-        # leak via a different error than an invalid token, so the
-        # require_admin_token call below produces the canonical 401/403.
-    _require_admin_token(
-        authorization, token,
-        # Bandit B106 false positive: this is the error message shown
-        # when the password env var is unset, not a hardcoded credential.
-        missing_password_detail=(  # nosec B106
-            "Match reports are disabled. Set OVERLAY_MANAGER_PASSWORD "
-            "for gated access or MATCH_REPORT_PUBLIC=true for open "
-            "access."
-        ),
-    )
-
-
-def _check_admin_access(authorization: str | None, token: str | None) -> None:
-    """Stricter sibling of :func:`_check_access` for destructive actions.
-
-    The public-mode shortcut from :func:`_check_access` deliberately does
-    not apply here: ``MATCH_REPORT_PUBLIC=true`` grants read-only access
-    (anyone with a non-guessable ``match_id`` can view a report), but
-    deletion must still go through the admin token. When
-    ``OVERLAY_MANAGER_PASSWORD`` is unset the operator has no way to
-    authenticate destructive calls — return 503 rather than silently
-    accepting them.
-    """
-    _require_admin_token(
-        authorization, token,
-        # Bandit B106 false positive: this is the error message shown
-        # when the password env var is unset, not a hardcoded credential.
-        missing_password_detail=(  # nosec B106
-            "Destructive match-archive actions are disabled. "
-            "Set OVERLAY_MANAGER_PASSWORD to enable them."
-        ),
-    )
-
-
-def _fmt_seconds(seconds: float | None) -> str:
-    if seconds is None:
-        return "—"
-    seconds = max(0, int(seconds))
-    h, rem = divmod(seconds, 3600)
-    m, s = divmod(rem, 60)
-    if h:
-        return f"{h}h {m:02d}m {s:02d}s"
-    return f"{m}m {s:02d}s"
-
-
-def _fmt_ts(ts: float | None) -> str:
-    if ts is None:
-        return "—"
-    try:
-        dt = datetime.datetime.fromtimestamp(float(ts), datetime.UTC)
-    except (TypeError, ValueError, OverflowError):
-        return "—"
-    return dt.strftime("%Y-%m-%d %H:%M UTC")
-
-
-_HEX_COLOR_RE = re.compile(r"^#(?:[0-9a-fA-F]{3}|[0-9a-fA-F]{6})$")
-
-
-def _team_color(customization: dict, team: int, primary: bool) -> str:
-    """Resolve a strict hex colour from the customization dict.
-
-    Falls back to a sensible default when the key is missing or the
-    stored value is anything but ``#RGB`` / ``#RRGGBB``. Strictness is
-    load-bearing: this value is interpolated into a CSS custom
-    property, so a malformed input could otherwise inject CSS.
-
-    Key priority: the team-identity colours (``Team 1 Color`` /
-    ``Team 1 Text Color``) come first because they're the *team's*
-    brand, set per-team in the operator UI. The overlay-wide
-    ``Color 1`` / ``Text Color 1`` keys are alternating row colours
-    and shouldn't override the team's own colour in the report.
-    """
-    fallback_bg = ("#0047AB", "#E21836")[team - 1]
-    fallback_fg = "#FFFFFF"
-    bg_keys = {
-        1: ("Team 1 Color", "Color 1", "color_primary"),
-        2: ("Team 2 Color", "Color 2", "color_primary"),
-    }
-    fg_keys = {
-        1: ("Team 1 Text Color", "Text Color 1"),
-        2: ("Team 2 Text Color", "Text Color 2"),
-    }
-    keys = bg_keys[team] if primary else fg_keys[team]
-    for key in keys:
-        value = customization.get(key)
-        if isinstance(value, str) and _HEX_COLOR_RE.match(value):
-            return value
-    return fallback_bg if primary else fallback_fg
-
-
-def _team_name(customization: dict, team: int) -> str:
-    # ``Team {n} Name`` is the canonical key the React control UI
-    # writes; ``Team {n} Text Name`` is the legacy alias the rest
-    # of the codebase still honours via ``Customization.A_TEAM`` /
-    # ``B_TEAM``. The overlays.uno cloud customization is also
-    # known to round-trip the legacy form depending on what the
-    # operator typed into the UNO panel — without the alias here
-    # the report falls back to the literal ``Team 1`` / ``Team 2``
-    # strings for any UNO-backed match. Snake-case and ``name{n}``
-    # cover older / external archives.
-    for key in (
-        f"Team {team} Name",
-        f"Team {team} Text Name",
-        f"team_{team}_name",
-        f"name{team}",
-    ):
-        value = customization.get(key)
-        if isinstance(value, str) and value.strip():
-            return value.strip()
-    return f"Team {team}"
-
-
-def _action_label(record: dict, locale: str) -> str:
-    """Human-readable label for an audit-log row in the report.
-
-    No ``(undone)`` suffix is emitted: ``_collapse_undos`` strips
-    every undo record (and its forward pair) before this function
-    runs, so a row reaching here always represents an action that
-    survived to the final state.
-    """
-    action = record.get("action", "")
-    params = record.get("params") or {}
-    team = params.get("team")
-    if action == "add_point":
-        return _t(locale, "actionPoint", team=team)
-    if action == "add_set":
-        return _t(locale, "actionSet", team=team)
-    if action == "add_timeout":
-        return _t(locale, "actionTimeout", team=team)
-    if action == "change_serve":
-        return _t(locale, "actionServe", team=team)
-    if action == "set_score":
-        return _t(
-            locale, "actionScore",
-            team=team,
-            set=params.get("set_number"),
-            value=params.get("value"),
-        )
-    if action == "reset":
-        return _t(locale, "actionReset")
-    return action or _t(locale, "actionUnknown")
-
-
-# Single source of truth for the timeline chip palette. Each entry
-# pairs the chip kind (used as a CSS modifier ``chip-{kind}``) with
-# the glyph rendered inside the accent strip and, when applicable,
-# the i18n key that names it in the bottom legend.
-#
-# Order matters — the legend section iterates the catalogue in the
-# order entries appear here, so the operator scans them top-to-
-# bottom (per-team points, then set / timeout / serve / edit /
-# reset).
-#
-# There is intentionally no ``undone`` entry: ``_collapse_undos``
-# strips both halves of every undo pair upstream, so no row that
-# reaches the timeline carries an undone state. The frontend
-# audit drawer's ``chipCatalogue.ts`` keeps an ``undone`` entry
-# because the live operator transcript still surfaces individual
-# undo records as their own rows.
-_CHIP_CATALOGUE: dict[str, dict[str, str | None]] = {
-    "point-t1": {"glyph": "+1", "legend_key": "legendPointT1"},
-    "point-t2": {"glyph": "+1", "legend_key": "legendPointT2"},
-    # Generic point chip used for legacy/missing-team rows. Not
-    # surfaced in the legend because the per-team variants already
-    # cover the shared semantics.
-    "point":    {"glyph": "+1", "legend_key": None},
-    "set":      {"glyph": "🏆", "legend_key": "legendSet"},
-    "timeout":  {"glyph": "⏸", "legend_key": "legendTimeout"},
-    "serve":    {"glyph": "⇄", "legend_key": "legendServe"},
-    "edit":     {"glyph": "✎", "legend_key": "legendEdit"},
-    "reset":    {"glyph": "⟲", "legend_key": "legendReset"},
-    # Final fallback — keeps unknown actions from rendering a blank
-    # accent strip. Intentionally unlabelled in the legend.
-    "other":    {"glyph": "•", "legend_key": None},
-}
-
-
-def _chip_glyph(kind: str) -> str:
-    """Glyph shown inside the chip accent strip for a given kind."""
-    entry = _CHIP_CATALOGUE.get(kind, _CHIP_CATALOGUE["other"])
-    glyph = entry["glyph"]
-    return glyph if isinstance(glyph, str) else "•"
-
-
-# Classifier-driven chip metadata. Returns the (modifier, glyph) pair
-# the timeline ``<li>`` and its accent strip use to differentiate
-# action kinds at a glance. Glyphs come from ``_CHIP_CATALOGUE`` so
-# the legend, the per-row strip and any future surface that needs
-# the same palette stay consistent without manual sync.
-def _chip_classifier(action: str, team: object) -> tuple[str, str]:
-    """Map an audit-record action+team to a chip ``(modifier, glyph)``.
-
-    ``modifier`` keys the chip's CSS class (``chip-{modifier}``) so
-    the stylesheet can paint a different accent and background per
-    action kind. Team-bound rows use ``point-t1`` / ``point-t2`` so
-    the running score reads alongside its team colour without
-    requiring per-team chip glyphs. ``_collapse_undos`` upstream
-    guarantees no undone records reach this function.
-    """
-    if action == "add_point":
-        if team == 1:
-            kind = "point-t1"
-        elif team == 2:
-            kind = "point-t2"
-        else:
-            kind = "point"
-    elif action == "add_set":
-        kind = "set"
-    elif action == "add_timeout":
-        kind = "timeout"
-    elif action == "change_serve":
-        kind = "serve"
-    elif action == "set_score":
-        kind = "edit"
-    elif action == "reset":
-        kind = "reset"
-    else:
-        kind = "other"
-    return (kind, _chip_glyph(kind))
-
-
-# ---------------------------------------------------------------------------
-# Audit-log derived helpers (running score, undo collapse, stats, charts)
-# ---------------------------------------------------------------------------
-
-def _coerce_int(raw: object) -> int | None:
-    """Best-effort ``int`` parse for audit-record numeric fields.
-
-    Accepts a real ``int`` directly or a string of digits (some
-    archive paths stringify scores). Returns ``None`` for anything
-    else. Caller decides whether to apply a positivity filter.
-    """
-    if isinstance(raw, int):
-        return raw
-    if isinstance(raw, str) and raw.isdigit():
-        return int(raw)
-    return None
-
-
-@overload
-def _safe_int(value: object) -> int | None: ...
-@overload
-def _safe_int(value: object, default: int) -> int: ...
-@overload
-def _safe_int(value: object, default: None) -> int | None: ...
-
-
-def _safe_int(value: object, default: int | None = None) -> int | None:
-    """Lenient ``int`` parse: like ``int(value)`` but returns *default* on failure.
-
-    Stricter callers should prefer :func:`_coerce_int`; this helper
-    mirrors the inline ``try: int(value) except (TypeError, ValueError)``
-    pattern that recurs across set/score parsing.
-    """
-    try:
-        return int(value)  # type: ignore[call-overload]
-    except (TypeError, ValueError):
-        return default
-
-
-def _result_score(record: dict, team: int) -> int | None:
-    """Pull the post-action score for *team* out of an audit ``result`` blob."""
-    block = (record.get("result") or {}).get(f"team_{team}") or {}
-    return _coerce_int(block.get("score"))
-
-
-def _result_set(record: dict) -> int | None:
-    """Set number this audit record applies to (1-indexed).
-
-    Prefers ``result.score_set`` when present: a set-winning
-    ``add_point`` advances ``current_set`` to the next set, but the
-    scores in the record (e.g. 25-23) belong to the *previous* set,
-    and ``score_set`` tags that explicitly. Falls back to
-    ``current_set`` for older audit records that predate the
-    ``score_set`` field.
-    """
-    result = record.get("result") or {}
-    for key in ("score_set", "current_set"):
-        n = _coerce_int(result.get(key))
-        if n is not None and n > 0:
-            return n
-    return None
-
-
-def _is_score_action(record: dict) -> bool:
-    return record.get("action") in ("add_point", "set_score")
-
-
-def _first_scoring_index(audit: list[dict]) -> int | None:
-    """Index of the first non-undo ``add_point`` / ``set_score`` record.
-
-    The audit log starts with whatever the operator did first — often a
-    ``reset`` to clear the previous match's state — but a "match" only
-    really begins once someone scores. Until the UI exposes a dedicated
-    "match start" marker, every report-side time anchor (relative
-    timestamps, duration, set durations, stats) snaps to this index.
-    Returns ``None`` when the audit has no scoring action at all.
-    """
-    for index, record in enumerate(audit):
-        if (record.get("params") or {}).get("undo"):
-            continue
-        if _is_score_action(record):
-            return index
-    return None
-
-
-def _trim_pregame(audit: list[dict]) -> list[dict]:
-    """Drop pre-first-scoring records (``reset``, stray timeouts, …).
-
-    Keeps everything from the first scoring action onward. When the
-    audit has no scoring action at all we return ``[]`` — the timeline
-    renderer already understands the empty case and shows "no audit
-    records", and falling through to the unfiltered list would let the
-    pregame noise leak back into the report.
-    """
-    idx = _first_scoring_index(audit)
-    return audit[idx:] if idx is not None else []
-
-
-def _played_set_count(final_state: dict, fallback: int) -> int:
-    """Highest set N with non-zero scoring data, clamped to *fallback*.
-
-    The match-history archive bundles ``team_X.scores.set_N`` for every
-    set up to ``sets_limit``, even sets that were never played (best
-    of 3 ending 2-0 still has ``set_3``: 0/0). Render only the sets
-    that actually saw points so empty trailing columns don't dilute
-    the report. When no set has scores yet (fresh archive) we collapse
-    to a single set frame rather than painting all ``sets_limit``
-    columns full of ``—``s.
-
-    *fallback* is honoured only as an upper bound — corrupt snapshots
-    reporting set N > sets_limit shouldn't paint a column the rules
-    don't allow.
-    """
-    teams = (final_state.get("team_1") or {}, final_state.get("team_2") or {})
-    highest = 0
-    for team in teams:
-        scores = team.get("scores") or {}
-        for key, value in scores.items():
-            if not isinstance(key, str) or not key.startswith("set_"):
-                continue
-            n = _safe_int(key.removeprefix("set_"))
-            if n is None:
-                continue
-            v = 0 if value is None else _safe_int(value, 0)
-            if v > 0:
-                highest = max(highest, n)
-    if highest == 0:
-        return 1
-    return min(highest, fallback) if fallback > 0 else highest
-
-
-def _collapse_undos(audit: list[dict]) -> list[dict]:
-    """Drop both halves of every ``undo`` pair from the rendered
-    timeline so undone actions never appear in the report.
-
-    Two cases reach this function:
-
-    * Live unified-undo logs (the common case), where
-      ``action_log.pop_last_forward`` already removed the original
-      forward physically and the audit-log just carries the
-      trailing undo record. The orphan undo is dropped because
-      the action it referenced no longer exists.
-    * Legacy / archived audit logs that still hold both a forward
-      record and the explicit undo that reversed it — typically
-      from pre-unification snapshots or replay-style fixtures. We
-      walk back to the most recent matching forward by
-      ``(action, team)`` and remove **both** the forward and the
-      undo, mirroring the live behaviour.
-
-    Net result: the report renders "as if the undone action never
-    happened". State-level aggregates already reflect the inverse
-    (the score / set / timeout counters never recorded the popped
-    increment), so the timeline can stay equally clean.
-    """
-    out: list[dict] = []
-    for record in audit:
-        params = record.get("params") or {}
-        if not params.get("undo"):
-            out.append(dict(record))
-            continue
-        # Walk back for the most recent forward record with the same
-        # ``(action, team)`` and remove it. The undo itself never
-        # reaches the output either — both halves disappear.
-        action = record.get("action")
-        team = params.get("team")
-        for index in range(len(out) - 1, -1, -1):
-            prior = out[index]
-            prior_params = prior.get("params") or {}
-            if (
-                prior.get("action") == action
-                and prior_params.get("team") == team
-                and not prior_params.get("undo")
-            ):
-                del out[index]
-                break
-        # No matching forward → orphan undo. Already not appended
-        # above; nothing else to do.
-    return out
-
-
-def _set_durations_from_audit(audit: list[dict]) -> dict[int, float]:
-    """Per-set duration in seconds derived from audit timestamps.
-
-    For each set we take ``max_ts - min_ts`` over records whose
-    ``current_set`` matches. Sets with fewer than two timestamps fall
-    out of the result map (no meaningful duration).
-    """
-    by_set: dict[int, list[float]] = {}
-    for record in audit:
-        if record.get("params", {}).get("undo"):
-            continue
-        set_num = _result_set(record)
-        ts = record.get("ts")
-        if set_num is None or not isinstance(ts, (int, float)):
-            continue
-        by_set.setdefault(set_num, []).append(float(ts))
-    durations: dict[int, float] = {}
-    for set_num, stamps in by_set.items():
-        if len(stamps) < 2:
-            continue
-        durations[set_num] = max(stamps) - min(stamps)
-    return durations
-
-
-def _running_score_pair(record: dict) -> tuple[int, int] | None:
-    """``(team1, team2)`` running score after this audit record, if known."""
-    s1 = _result_score(record, 1)
-    s2 = _result_score(record, 2)
-    if s1 is None or s2 is None:
-        return None
-    return (s1, s2)
-
-
-def _format_relative_ts(ts: float | None, base_ts: float | None) -> str:
-    """``+0:15`` / ``+1:23:45`` style offset from the match start.
-
-    Returns ``"—"`` for missing inputs and ``"+0:00"`` for the very
-    first record (relative to itself). Negative deltas would mean an
-    out-of-order log; we clamp to zero rather than render a minus sign
-    that would baffle the operator.
-    """
-    if ts is None or base_ts is None:
-        return "—"
-    delta = max(0, int(float(ts) - float(base_ts)))
-    h, rem = divmod(delta, 3600)
-    m, s = divmod(rem, 60)
-    if h:
-        return f"+{h}:{m:02d}:{s:02d}"
-    return f"+{m}:{s:02d}"
-
-
-def _logo_url(customization: dict, team: int) -> str | None:
-    """Return a sanitised logo URL for *team*, or ``None`` if missing.
-
-    Only ``http(s):`` and ``data:`` schemes are accepted — the URL is
-    interpolated into ``<img src=…>`` and any other scheme would invite
-    XSS via ``javascript:``-style payloads.
-    """
-    for key in (f"Team {team} Logo", f"team_{team}_logo", f"logo{team}"):
-        value = customization.get(key)
-        if not isinstance(value, str):
-            continue
-        candidate = value.strip()
-        if not candidate:
-            continue
-        lowered = candidate.lower()
-        if lowered.startswith(("http://", "https://", "data:image/")):
-            return candidate
-    return None
-
-
-def _timeouts_per_set(audit: list[dict]) -> dict[int, dict[int, int]]:
-    """Count ``add_timeout`` actions per (set, team) from the audit log.
-
-    The live snapshot's ``team_X.timeouts`` is the current set's count,
-    so reconstructing the full per-set table for the printed report
-    requires walking the audit log (each ``add_timeout`` record
-    carries the set it landed in via ``result.current_set``). Returns
-    ``{set_num: {team: count}}``; callers consult
-    ``.get(set, {}).get(team, 0)`` for safe lookup.
-    """
-    out: dict[int, dict[int, int]] = {}
-    for record in audit:
-        if record.get("action") != "add_timeout":
-            continue
-        params = record.get("params") or {}
-        if params.get("undo"):
-            continue
-        team = params.get("team")
-        set_num = _result_set(record)
-        if team not in (1, 2) or set_num is None:
-            continue
-        out.setdefault(set_num, {}).setdefault(team, 0)
-        out[set_num][team] += 1
-    return out
-
-
-def _compute_stats(audit: list[dict]) -> dict:
-    """Compute the Highlights block (longest streak, biggest comeback, totals).
-
-    All metrics derive purely from the audit log so the report stays
-    consistent with the "scoring trajectory" the audit promises. Set
-    points are scored per-set. Comebacks are tracked per-team and
-    split into two flavours:
-
-    * ``set_win``: the largest deficit a team erased *and* went on to
-      win the set with — i.e. the deficit faced by the eventual set
-      winner.
-    * ``partial``: the largest deficit reduction (peak deficit minus
-      the smallest subsequent deficit) achieved by a team that
-      ultimately *lost* the set. This surfaces near-comebacks that
-      didn't quite finish.
-
-    Storing both per team lets the renderer detect a tie when both
-    sides happen to share the same maximum.
-    """
-    longest_streak: dict[str, Any] = {"team": None, "n": 0, "set": None}
-    # Per-team peak deficit erased by the eventual set winner.
-    set_win_comeback: dict[int, dict] = {
-        1: {"deficit": 0, "set": None},
-        2: {"deficit": 0, "set": None},
-    }
-    # Per-team peak deficit reduction by a losing team (partial recovery).
-    partial_comeback: dict[int, dict] = {
-        1: {"deficit": 0, "set": None},
-        2: {"deficit": 0, "set": None},
-    }
-    longest_rally: dict[str, Any] = {"duration_s": 0.0, "set": None}
-    total_points = 0
-    # Per-team total of won rallies (every ``add_point`` for the team,
-    # tagged or not), used as the denominator for the point-composition
-    # percentages in the report.
-    total_points_by_team: dict[int, int] = {1: 0, 2: 0}
-    # Optional per-point classification tallies (opt-in scouting tags).
-    # ``point_types`` counts every tagged forward point per team;
-    # ``error_types`` breaks the ``opp_error`` total into its specific
-    # causes (a subset of opp_error — untyped opp_errors aren't here).
-    point_types: dict[int, dict[str, int]] = {
-        1: dict.fromkeys(POINT_TYPES, 0),
-        2: dict.fromkeys(POINT_TYPES, 0),
-    }
-    error_types: dict[int, dict[str, int]] = {
-        1: dict.fromkeys(ERROR_TYPES, 0),
-        2: dict.fromkeys(ERROR_TYPES, 0),
-    }
-    # Per-set point-type tallies, so the set-summary recap can scope the
-    # breakdown to the displayed set (same pattern as the per-set streak
-    # / services variants). Only sets that have a tagged point appear.
-    point_types_by_set: dict[int, dict[int, dict[str, int]]] = {}
-
-    by_set: dict[int, list[dict]] = {}
-    for record in audit:
-        if record.get("params", {}).get("undo"):
-            continue
-        set_num = _result_set(record)
-        if set_num is None:
-            continue
-        by_set.setdefault(set_num, []).append(record)
-
-    for set_num, set_records in by_set.items():
-        # Longest streak: count consecutive ``add_point`` actions by
-        # the same team. Manual set_score breaks the streak.
-        streak_team: int | None = None
-        streak_n = 0
-        for r in set_records:
-            if r.get("action") != "add_point":
-                streak_team, streak_n = None, 0
-                continue
-            team = (r.get("params") or {}).get("team")
-            if team == streak_team:
-                streak_n += 1
-            else:
-                streak_team, streak_n = team, 1
-            if streak_n > longest_streak["n"]:
-                longest_streak = {"team": team, "n": streak_n, "set": set_num}
-            total_points += 1
-            if team in (1, 2):
-                total_points_by_team[team] += 1
-                params = r.get("params") or {}
-                pt = params.get("point_type")
-                if pt in point_types[team]:
-                    point_types[team][pt] += 1
-                    pts_set = point_types_by_set.setdefault(
-                        set_num,
-                        {
-                            1: dict.fromkeys(POINT_TYPES, 0),
-                            2: dict.fromkeys(POINT_TYPES, 0),
-                        },
-                    )
-                    pts_set[team][pt] += 1
-                    if pt == "opp_error":
-                        et = params.get("error_type")
-                        if et in error_types[team]:
-                            error_types[team][et] += 1
-
-        # Comebacks. Walk the running scores once, then split:
-        #   * winner's peak deficit  → set_win comeback for the winner
-        #   * loser's peak deficit minus the smallest subsequent deficit
-        #     → partial comeback for the loser (they trimmed the gap
-        #     but didn't close it).
-        if not set_records:
-            continue
-        last_score = _running_score_pair(set_records[-1])
-        if not last_score:
-            continue
-        winner = 1 if last_score[0] > last_score[1] else 2
-        loser = 2 if winner == 1 else 1
-        winner_peak_deficit = 0
-        loser_peak_deficit = 0
-        loser_min_after_peak = 0
-        loser_max_recovery = 0
-        for r in set_records:
-            pair = _running_score_pair(r)
-            if not pair:
-                continue
-            t1, t2 = pair
-            winner_deficit = (t2 - t1) if winner == 1 else (t1 - t2)
-            loser_deficit = -winner_deficit
-            if winner_deficit > winner_peak_deficit:
-                winner_peak_deficit = winner_deficit
-            if loser_deficit > loser_peak_deficit:
-                loser_peak_deficit = loser_deficit
-                loser_min_after_peak = loser_deficit
-            elif loser_peak_deficit > 0:
-                # Only count "recovery" once the loser has actually
-                # trailed. Otherwise a team that led 5-0 from the
-                # start and then collapsed would post a phantom
-                # 5-point partial comeback — they never erased
-                # anything, they just bled their early lead.
-                # Clamp at 0: the comeback ends at the tie. A loser
-                # who briefly took a lead mid-set still gets credit
-                # for "points recovered while trailing" up to the
-                # tying point, but the subsequent lead points are a
-                # separate (and short-lived) story.
-                clamped = max(0, loser_deficit)
-                if clamped < loser_min_after_peak:
-                    loser_min_after_peak = clamped
-                    recovery = loser_peak_deficit - loser_min_after_peak
-                    if recovery > loser_max_recovery:
-                        loser_max_recovery = recovery
-        if winner_peak_deficit > set_win_comeback[winner]["deficit"]:
-            set_win_comeback[winner] = {
-                "deficit": winner_peak_deficit, "set": set_num,
-            }
-        if loser_max_recovery > partial_comeback[loser]["deficit"]:
-            partial_comeback[loser] = {
-                "deficit": loser_max_recovery, "set": set_num,
-            }
-
-        # Longest rally: the gap between consecutive scoring actions
-        # within the set is a proxy for "longest point" — without
-        # ball-in-play instrumentation that's the closest the audit
-        # log can give us. Restrict to ``add_point`` only: a manual
-        # ``set_score`` override is an editorial action (operator
-        # correcting a score after the fact) and including it would
-        # report the *editing* delay as a rally. The audit log is
-        # append-only so ``set_records`` is already in chronological
-        # order — no extra sort needed.
-        scoring_ts: list[float] = []
-        for r in set_records:
-            if r.get("action") != "add_point":
-                continue
-            ts = r.get("ts")
-            if isinstance(ts, (int, float)):
-                scoring_ts.append(float(ts))
-        for i in range(1, len(scoring_ts)):
-            delta = scoring_ts[i] - scoring_ts[i - 1]
-            if delta > longest_rally["duration_s"]:
-                longest_rally = {"duration_s": delta, "set": set_num}
-
-    return {
-        "longest_streak": longest_streak,
-        "set_win_comeback": set_win_comeback,
-        "partial_comeback": partial_comeback,
-        "longest_rally": longest_rally,
-        "total_points": total_points,
-        "total_points_by_team": total_points_by_team,
-        "set_durations": _set_durations_from_audit(audit),
-        "point_types": point_types,
-        "error_types": error_types,
-        "point_types_by_set": point_types_by_set,
-    }
-
-
-# ---------------------------------------------------------------------------
-# Contrast-safe colour pickers for the chart / highlight surfaces
-# ---------------------------------------------------------------------------
-
-# Distinguishable fallback palette for teams whose own brand colour is
-# either too light to read on the report's white surface or
-# indistinguishable from the other team's colour. ``team_index`` is
-# 0-based so callers can use ``[team-1]``.
-_CHART_FALLBACK = ("#0047AB", "#E21836")
-# Luminance threshold above which the colour is "too light" for the
-# white-ish report background. Computed as relative luminance per
-# WCAG so 0.0 == black, 1.0 == white. ~0.85 keeps pastels usable but
-# rejects pure white / very light yellows.
-_LIGHTNESS_REJECT = 0.85
-
-
-def _hex_to_rgb(hex_color: str) -> tuple[int, int, int] | None:
-    """Parse ``#RGB`` / ``#RRGGBB`` into ``(r, g, b)`` ∈ [0, 255]."""
-    if not isinstance(hex_color, str) or not _HEX_COLOR_RE.match(hex_color):
-        return None
-    body = hex_color.lstrip("#")
-    if len(body) == 3:
-        body = "".join(ch * 2 for ch in body)
-    return (int(body[0:2], 16), int(body[2:4], 16), int(body[4:6], 16))
-
-
-def _relative_luminance(hex_color: str) -> float | None:
-    """WCAG relative luminance for *hex_color*, or ``None`` on failure."""
-    rgb = _hex_to_rgb(hex_color)
-    if rgb is None:
-        return None
-
-    def _channel(v: int) -> float:
-        c = v / 255.0
-        return c / 12.92 if c <= 0.03928 else ((c + 0.055) / 1.055) ** 2.4
-
-    r, g, b = (_channel(c) for c in rgb)
-    return 0.2126 * r + 0.7152 * g + 0.0722 * b
-
-
-def _chart_color(team: int, primary: str, fg: str) -> str:
-    """Pick a chart-/highlight-safe colour for *team*.
-
-    The team's primary brand colour is used when it has acceptable
-    contrast against the report's white background. Otherwise we
-    fall back to the team's text colour (which is, by design,
-    high-contrast against the brand colour and usually against
-    the page too) and finally to a fixed palette. This keeps every
-    rendered datapoint visible without losing team identity.
-    """
-    candidates = [primary, fg, _CHART_FALLBACK[(team - 1) % 2]]
-    for candidate in candidates:
-        lum = _relative_luminance(candidate)
-        if lum is not None and lum < _LIGHTNESS_REJECT:
-            return candidate
-    return _CHART_FALLBACK[(team - 1) % 2]
-
-
-def _ensure_distinct_chart_colors(c1: str, c2: str) -> tuple[str, str]:
-    """If both teams resolve to the same chart colour, force team 2 to fallback."""
-    if c1.lower() != c2.lower():
-        return c1, c2
-    fallback = _CHART_FALLBACK[1] if c1.lower() != _CHART_FALLBACK[1].lower() \
-        else _CHART_FALLBACK[0]
-    return c1, fallback
-
-
-# ---------------------------------------------------------------------------
-# Original ``_compute_stats`` continues here (kept logically below the
-# helpers above so the highlight / chart block has the colour-safety
-# tools imported by name).
-# ---------------------------------------------------------------------------
-
-
-# Anything beyond this gap between consecutive points is treated as
-# "the operator left the tab open / time isn't trustworthy", and the
-# chart falls back to the rally-number X-axis. 15 minutes is well
-# beyond a normal set-break and well short of operator-distraction
-# territory.
-_TIME_AXIS_MAX_GAP_S = 15 * 60
-
-
-def _format_mmss(seconds: float) -> str:
-    """``MM:SS`` (no leading zero on minutes) for the X-axis label."""
-    seconds = max(0, int(seconds))
-    m, s = divmod(seconds, 60)
-    return f"{m}:{s:02d}"
-
-
-def _render_score_chart(
-    set_records: list[dict], *,
-    t1_color: str, t2_color: str,
-    timeouts: list[dict] | None = None,
-) -> str:
-    """Inline SVG showing how each team's score evolved through a set.
-
-    X axis: time elapsed since the first point of the set (``MM:SS``)
-    when the audit timestamps look reliable. If any gap between two
-    consecutive scoring records exceeds :data:`_TIME_AXIS_MAX_GAP_S`
-    we fall back to plain rally-number indexing — that's the signal
-    the operator stepped away and the timestamps stopped reflecting
-    play. Y axis: points scored, labelled 0 / mid / max. Each rally
-    datapoint is marked with a small filled circle so single-point
-    spikes are legible. One polyline per team, coloured via the
-    contrast-safe palette resolved upstream. Pure SVG, no JS, so it
-    survives "Save as PDF".
-
-    *timeouts* (optional) is the list of ``add_timeout`` audit
-    records belonging to this set, already collapsed for undos by
-    the caller. Each one renders as a thin dashed vertical guide
-    line in the calling team's colour, with a small downward
-    triangle perched above the chart so the operator can correlate
-    "score stalled here" with "the team called timeout".
-
-    Returns a placeholder string when the set has fewer than two
-    scoring records (nothing to plot).
-    """
-    points: list[tuple[int, int]] = []
-    timestamps: list[float | None] = []
-    for r in set_records:
-        pair = _running_score_pair(r)
-        if not pair:
-            continue
-        points.append(pair)
-        ts = r.get("ts")
-        timestamps.append(float(ts) if isinstance(ts, (int, float)) else None)
-    if len(points) < 2:
-        return ""
-
-    max_score = max(max(p) for p in points)
-    width, height = 360, 150
-    pad_x_left, pad_x_right = 32, 18
-    pad_y_top, pad_y_bottom = 14, 26
-    plot_w = width - pad_x_left - pad_x_right
-    plot_h = height - pad_y_top - pad_y_bottom
-    last_idx = len(points) - 1
-    if last_idx == 0 or max_score == 0:
-        return ""
-
-    # Decide axis mode. Time mode requires every record to carry a
-    # timestamp *and* no gap between consecutive records to exceed
-    # the trust threshold — anything bigger means the operator was
-    # AFK and the wallclock no longer tracks play, so we fall back
-    # to rally-number indexing rather than compress the whole set
-    # into a thin slice on the left.
-    times: list[float] | None = None
-    if all(t is not None for t in timestamps):
-        # ``timestamps`` is structurally ``list[Optional[float]]``;
-        # the ``all(...)`` check above narrows it but mypy can't see
-        # through it, so build the float-only list explicitly.
-        times = [float(t) for t in timestamps if t is not None]
-        for i in range(1, len(times)):
-            gap = times[i] - times[i - 1]
-            # ``gap > threshold``: operator was AFK, wallclock no
-            # longer tracks play. ``gap < 0``: non-monotonic
-            # timestamps (clock skew, NTP correction) — would
-            # otherwise plot points outside the SVG viewport. Both
-            # downgrade to the rally-number fallback.
-            if gap > _TIME_AXIS_MAX_GAP_S or gap < 0:
-                times = None
-                break
-
-    use_time_axis = times is not None
-    if times is not None:
-        base_ts = times[0]
-        x_values: list[float] = [t - base_ts for t in times]
-        # Guard against a degenerate "all points at the same ts" set:
-        # the polyline would collapse, but we still need a non-zero
-        # divisor for the projection.
-        x_max = x_values[-1] if x_values[-1] > 0 else 1.0
-    else:
-        x_values = [float(i) for i in range(len(points))]
-        x_max = float(last_idx) if last_idx else 1.0
-
-    def _project(idx: int, score: int) -> tuple[float, float]:
-        x_norm = x_values[idx] / x_max if x_max else 0.0
-        x = pad_x_left + x_norm * plot_w
-        y = pad_y_top + plot_h - (score / max_score) * plot_h
-        return x, y
-
-    mid_score = max_score // 2 if max_score >= 2 else max_score
-    y_ticks = sorted({0, mid_score, max_score})
-
-    grid_lines = "".join(
-        f'<line x1="{pad_x_left}" y1="{_project(0, v)[1]:.1f}" '
-        f'x2="{pad_x_left + plot_w}" y2="{_project(0, v)[1]:.1f}" '
-        f'stroke="#e0e0e0" stroke-width="1" stroke-dasharray="2,3" />'
-        for v in y_ticks
-    )
-
-    y_labels = "".join(
-        f'<text x="{pad_x_left - 4}" y="{_project(0, v)[1] + 3:.1f}" '
-        f'text-anchor="end" font-size="9" fill="#999">{v}</text>'
-        for v in y_ticks
-    )
-
-    if use_time_axis:
-        # Endpoints: ``0:00`` → ``M:SS`` of the last rally relative
-        # to the set's first point.
-        left_label = "0:00"
-        right_label = _format_mmss(x_values[-1])
-    else:
-        # 1-indexed rally numbers, matching prior behaviour.
-        left_label = "1"
-        right_label = str(len(points))
-
-    x_labels = (
-        f'<text x="{pad_x_left}" y="{height - 8}" text-anchor="start" '
-        f'font-size="9" fill="#999">{html.escape(left_label)}</text>'
-        f'<text x="{pad_x_left + plot_w}" y="{height - 8}" '
-        f'text-anchor="end" font-size="9" fill="#999">{html.escape(right_label)}</text>'
-    )
-
-    def _polyline(team_idx: int, color: str) -> str:
-        coords = " ".join(
-            f"{x:.1f},{y:.1f}"
-            for x, y in (_project(i, p[team_idx]) for i, p in enumerate(points))
-        )
-        return (
-            f'<polyline fill="none" stroke="{html.escape(color)}" '
-            f'stroke-width="2" points="{coords}" />'
-        )
-
-    def _markers(team_idx: int, color: str) -> str:
-        # ``r=2.5`` so they sit just above polyline thickness — big
-        # enough to read on print, small enough not to obscure rapid
-        # back-and-forth swings in volleyball-style 25-point sets.
-        return "".join(
-            f'<circle cx="{x:.1f}" cy="{y:.1f}" r="2.5" '
-            f'fill="{html.escape(color)}" />'
-            for x, y in (_project(i, p[team_idx]) for i, p in enumerate(points))
-        )
-
-    def _timeout_markers() -> str:
-        if not timeouts:
-            return ""
-        # Reuse ``timestamps`` (1:1 with the plotted ``points``)
-        # rather than rebuilding from ``set_records`` — the polyline
-        # already filters out records that lack a ``_running_score_pair``,
-        # so a parallel rebuild here would drift the rally indices.
-        # Some entries may still be ``None`` when the chart fell back
-        # to rally mode because of missing timestamps; those score
-        # records simply don't contribute to the timeout's rally
-        # position.
-        items: list[str] = []
-        for record in timeouts:
-            params = record.get("params") or {}
-            team = params.get("team")
-            if team not in (1, 2):
-                continue
-            ts = record.get("ts")
-            if not isinstance(ts, (int, float)):
-                continue
-            ts_float = float(ts)
-            if use_time_axis and times is not None:
-                # Time mode: anchor to the same ``base_ts`` the score
-                # polyline uses. A timeout called *before* the first
-                # point of the set lands at x=0 (left edge) rather
-                # than off-canvas.
-                x_val = max(0.0, ts_float - times[0])
-                x_val = min(x_val, x_max)
-            else:
-                # Rally mode: count plotted score records whose
-                # timestamp is ``<= ts_float``. ``idx`` is the rally
-                # number after which the timeout was called; clamp
-                # into ``[0, last_idx]`` so a post-final-point
-                # timeout still renders on the right edge. Skip
-                # ``None`` entries (records without a timestamp)
-                # rather than letting the comparison raise.
-                idx = sum(
-                    1 for t in timestamps if t is not None and t <= ts_float
-                ) - 1
-                idx = max(0, min(idx, last_idx))
-                x_val = float(idx)
-            x_norm = x_val / x_max if x_max else 0.0
-            x = pad_x_left + x_norm * plot_w
-            color = t1_color if team == 1 else t2_color
-            color_safe = html.escape(color)
-            # Dashed guide spanning the plot height + a small clock
-            # face perched ~9 px above the plot area so it doesn't
-            # collide with the polylines or the ``max_score`` Y label.
-            # The glyph is a hand-rolled inline SVG (circle + two
-            # hands) rather than a Material Icons font ref so the
-            # report keeps surviving "Save as PDF" with no external
-            # font load.
-            cy = pad_y_top - 5
-            items.append(
-                f'<line class="set-chart-timeout" data-team="{team}" '
-                f'x1="{x:.1f}" y1="{pad_y_top:.1f}" '
-                f'x2="{x:.1f}" y2="{pad_y_top + plot_h:.1f}" '
-                f'stroke="{color_safe}" stroke-width="1" '
-                f'stroke-dasharray="3,3" opacity="0.55" />'
-                # Lift the shared stroke attributes to the ``<g>`` so
-                # the children only override what they need (the face
-                # gets a slightly thicker border than the hands).
-                f'<g class="set-chart-timeout-glyph" data-team="{team}" '
-                f'stroke="{color_safe}" stroke-width="1" '
-                f'stroke-linecap="round">'
-                f'<circle cx="{x:.1f}" cy="{cy:.1f}" r="3.5" '
-                f'fill="none" stroke-width="1.2" />'
-                # Minute hand: pointing up to "12".
-                f'<line x1="{x:.1f}" y1="{cy:.1f}" '
-                f'x2="{x:.1f}" y2="{cy - 2.2:.1f}" />'
-                # Hour hand: pointing right to "3".
-                f'<line x1="{x:.1f}" y1="{cy:.1f}" '
-                f'x2="{x + 1.7:.1f}" y2="{cy:.1f}" />'
-                f'</g>'
-            )
-        return "".join(items)
-
-    # ``data-x-axis`` lets tests assert which mode kicked in without
-    # parsing the rendered labels.
-    axis_attr = "time" if use_time_axis else "rally"
-    return (
-        f'<svg viewBox="0 0 {width} {height}" role="img" '
-        f'class="set-chart" data-x-axis="{axis_attr}" '
-        f'preserveAspectRatio="xMidYMid meet">'
-        f'<rect x="0" y="0" width="{width}" height="{height}" '
-        f'fill="transparent" />'
-        f'{grid_lines}{y_labels}{x_labels}'
-        f'{_polyline(0, t1_color)}{_polyline(1, t2_color)}'
-        f'{_markers(0, t1_color)}{_markers(1, t2_color)}'
-        f'{_timeout_markers()}'
-        f'</svg>'
-    )
-
-
-def _render_set_durations_row(durations: dict[int, float], set_count: int) -> str:
-    cells = []
-    for i in range(1, set_count + 1):
-        if i in durations:
-            cells.append(f"<td>{html.escape(_fmt_seconds(durations[i]))}</td>")
-        else:
-            cells.append("<td>—</td>")
-    return "".join(cells)
-
-
-def _render_highlights(
-    stats: dict, locale: str,
-    *, team1_name: str, team2_name: str,
-) -> str:
-    """Build the Highlights grid (longest streak / comeback / totals / set durations).
-
-    *team1_name* / *team2_name* are the human-readable team labels so
-    the cards say "Alpha" instead of the cryptic "Team 1". Falls back
-    to the i18n ``team`` template when a card references a team
-    number we somehow can't map (defensive — shouldn't happen).
-    """
-    cards: list[str] = []
-
-    def _team_label(team: int | None) -> str:
-        if team == 1:
-            return team1_name
-        if team == 2:
-            return team2_name
-        if team:
-            return _t(locale, "team", team=team)
-        return ""
-
-    def _card(label: str, value: str, detail: str = "") -> None:
-        detail_html = (
-            f'<div class="detail">{html.escape(detail)}</div>' if detail else ""
-        )
-        cards.append(
-            f'<div class="highlight"><div class="label">{html.escape(label)}</div>'
-            f'<div class="value">{html.escape(value)}</div>{detail_html}</div>'
-        )
-
-    streak = stats.get("longest_streak") or {}
-    if streak.get("n", 0) >= 2 and streak.get("team"):
-        _card(
-            _t(locale, "highlightStreak"),
-            _t(locale, "pointsValue", n=streak["n"]),
-            f"{_team_label(streak['team'])} · "
-            + _t(locale, "setLabel", n=streak.get("set") or "?"),
-        )
-
-    # Set-winning comeback: surface only big erased deficits (≥5 pts).
-    # When both teams' best erased deficit is the same, render a single
-    # tied-card instead of arbitrarily picking one team.
-    set_win = stats.get("set_win_comeback") or {}
-    sw1 = (set_win.get(1) or {}).get("deficit", 0)
-    sw2 = (set_win.get(2) or {}).get("deficit", 0)
-    sw_max = max(sw1, sw2)
-    if sw_max >= 5:
-        if sw1 == sw2:
-            _card(
-                _t(locale, "highlightComeback"),
-                _t(locale, "pointsValue", n=sw_max),
-                _t(locale, "comebackTie"),
-            )
-        else:
-            winner_team = 1 if sw1 > sw2 else 2
-            entry = set_win[winner_team]
-            _card(
-                _t(locale, "highlightComeback"),
-                _t(locale, "deltaValue",
-                   n=entry["deficit"], set=entry.get("set") or "?"),
-                _team_label(winner_team),
-            )
-
-    # Partial comeback: a deficit a team trimmed but couldn't close.
-    # Threshold > 3 pts so we don't celebrate a one-rally swing.
-    partial = stats.get("partial_comeback") or {}
-    p1 = (partial.get(1) or {}).get("deficit", 0)
-    p2 = (partial.get(2) or {}).get("deficit", 0)
-    p_max = max(p1, p2)
-    if p_max > 3:
-        if p1 == p2:
-            _card(
-                _t(locale, "highlightPartialComeback"),
-                _t(locale, "pointsValue", n=p_max),
-                _t(locale, "comebackTie"),
-            )
-        else:
-            loser_team = 1 if p1 > p2 else 2
-            entry = partial[loser_team]
-            _card(
-                _t(locale, "highlightPartialComeback"),
-                _t(locale, "partialDeltaValue",
-                   n=entry["deficit"], set=entry.get("set") or "?"),
-                _team_label(loser_team),
-            )
-
-    rally = stats.get("longest_rally") or {}
-    rally_duration = rally.get("duration_s") or 0
-    if rally_duration >= 1 and rally.get("set"):
-        # Sub-second rallies are noise (back-to-back action_log
-        # appends at the same wallclock); only show when there's
-        # actually a measurable gap.
-        _card(
-            _t(locale, "highlightLongestRally"),
-            _fmt_seconds(rally_duration),
-            _t(locale, "setLabel", n=rally["set"]),
-        )
-
-    total = stats.get("total_points", 0)
-    if total:
-        _card(_t(locale, "highlightTotalPoints"), str(total))
-
-    durations = stats.get("set_durations") or {}
-    if durations:
-        longest = max(durations.items(), key=lambda kv: kv[1])
-        shortest = min(durations.items(), key=lambda kv: kv[1])
-        _card(
-            _t(locale, "highlightLongestSet"),
-            _fmt_seconds(longest[1]),
-            _t(locale, "setLabel", n=longest[0]),
-        )
-        if shortest[0] != longest[0]:
-            _card(
-                _t(locale, "highlightShortestSet"),
-                _fmt_seconds(shortest[1]),
-                _t(locale, "setLabel", n=shortest[0]),
-            )
-
-    # Per-point classification breakdown (opt-in scouting tags). One
-    # card per team that recorded at least one classified point; the
-    # value is the classified total and the detail spells out the mix,
-    # with opponent errors further broken down by cause when tracked.
-    point_types = stats.get("point_types") or {}
-    error_types = stats.get("error_types") or {}
-    _pt_label_keys = {
-        "ace": "pointTypeAce",
-        "kill": "pointTypeKill",
-        "block": "pointTypeBlock",
-        "opp_error": "pointTypeOppError",
-    }
-    _et_label_keys = {
-        "serve_error": "errorTypeServe",
-        "attack_error": "errorTypeAttack",
-        "reception_error": "errorTypeReception",
-        "ball_handling": "errorTypeBallHandling",
-        "net_fault": "errorTypeNet",
-        "position_fault": "errorTypePosition",
-        "other": "errorTypeOther",
-    }
-    totals_by_team = stats.get("total_points_by_team") or {}
-
-    def _pct(n: int, denom: int) -> int:
-        return round(100 * n / denom) if denom else 0
-
-    # Card 1 — point composition: how each team scored, each type as a
-    # share of that team's total points won (the remainder, if any, is
-    # untagged). Shown only for teams with at least one classified point.
-    for team in (1, 2):
-        counts = point_types.get(team) or {}
-        total_typed = sum(v for v in counts.values() if isinstance(v, int))
-        if total_typed <= 0:
-            continue
-        team_total = totals_by_team.get(team) or 0
-        parts = []
-        for k in ("ace", "kill", "block", "opp_error"):
-            n = counts.get(k) or 0
-            if not n:
-                continue
-            # "Label: N" (plural category label) reads grammatically at
-            # any count, unlike "N label" which yields "3 kill".
-            label = f"{_t(locale, _pt_label_keys[k])}: {n}"
-            if team_total:
-                label += f" ({_pct(n, team_total)}%)"
-            parts.append(label)
-        _card(
-            f"{_team_label(team)} · {_t(locale, 'pointTypesHeading')}",
-            str(total_typed),
-            " · ".join(parts),
-        )
-
-    # Card 2 — own errors: points a team gave away through its own faults,
-    # i.e. the opponent's ``opp_error`` tally (and its cause breakdown),
-    # plus the share of the opponent's points those mistakes accounted for.
-    for team in (1, 2):
-        opp = 2 if team == 1 else 1
-        gifted = (point_types.get(opp) or {}).get("opp_error") or 0
-        if gifted <= 0:
-            continue
-        opp_total = totals_by_team.get(opp) or 0
-        errs = error_types.get(opp) or {}
-        # "Label: N" (plural cause label) — grammatical at any count and
-        # self-describing, so no separate "errors:" lead-in is needed.
-        err_parts = [
-            f"{_t(locale, _et_label_keys[k])}: {errs[k]}"
-            for k in ERROR_TYPES
-            if errs.get(k)
-        ]
-        detail = (
-            _t(locale, "ownErrorsShare", pct=_pct(gifted, opp_total))
-            if opp_total else ""
-        )
-        if err_parts:
-            sep = " — " if detail else ""
-            detail += sep + " · ".join(err_parts)
-        _card(
-            f"{_team_label(team)} · {_t(locale, 'ownErrorsHeading')}",
-            str(gifted),
-            detail,
-        )
-
-    if not cards:
-        # Empty matches still render an explicit "no highlights" card
-        # rather than collapsing the whole section into a void.
-        _card(_t(locale, "highlights"), "—")
-    return "".join(cards)
-
-
-def _render_charts(
-    audit: list[dict], set_count: int, locale: str,
-    *, t1_name: str, t2_name: str, t1_color: str, t2_color: str,
-) -> str:
-    """Build the per-set score-evolution chart grid."""
-    scores_by_set: dict[int, list[dict]] = {}
-    timeouts_by_set: dict[int, list[dict]] = {}
-    for record in audit:
-        if record.get("params", {}).get("undo"):
-            continue
-        set_num = _result_set(record)
-        if set_num is None:
-            continue
-        if _is_score_action(record):
-            scores_by_set.setdefault(set_num, []).append(record)
-        elif record.get("action") == "add_timeout":
-            timeouts_by_set.setdefault(set_num, []).append(record)
-
-    timeout_legend = (
-        f'<span class="swatch swatch-timeout" aria-hidden="true"></span>'
-        f'{html.escape(_t(locale, "legendTimeout"))}'
-    )
-
-    cards: list[str] = []
-    for i in range(1, set_count + 1):
-        records = scores_by_set.get(i, [])
-        chart = _render_score_chart(
-            records,
-            t1_color=t1_color, t2_color=t2_color,
-            timeouts=timeouts_by_set.get(i, []),
-        )
-        body = chart or (
-            f'<p class="legend">{html.escape(_t(locale, "noScoreEvolution"))}</p>'
-        )
-        # Timeout swatch only when this set had at least one — the
-        # operator doesn't need a "Timeout" key on a clean set.
-        timeout_html = (
-            timeout_legend if timeouts_by_set.get(i) else ""
-        )
-        legend = (
-            f'<div class="legend">'
-            f'<span class="swatch" style="background: {html.escape(t1_color)};"></span>{html.escape(t1_name)}'
-            f'<span class="swatch" style="background: {html.escape(t2_color)};"></span>{html.escape(t2_name)}'
-            f'{timeout_html}'
-            f'</div>'
-        )
-        cards.append(
-            f'<div class="chart-card">'
-            f'<h3>{html.escape(_t(locale, "setLabel", n=i))}</h3>'
-            f'{legend}{body}'
-            f'</div>'
-        )
-    return "".join(cards) or ""
-
-
-def _render_timeline(
-    audit: list[dict], locale: str, set_count: int,
-    *, base_ts: float | None = None,
-) -> str:
-    """Group the audit by set and emit running-score-aware list items.
-
-    *audit* is expected to already be collapsed via
-    ``_collapse_undos`` upstream — every consumer in the report
-    pipeline shares the same collapsed slice so the timeline,
-    timeouts row, highlights and charts stay coherent.
-
-    *base_ts* is the explicit match-start anchor (Start-match
-    button or first-point auto-arm). When supplied, relative
-    timestamps are measured from there — so a point scored 5 min
-    after Start-match reads ``+5:00``, not ``+0:00``. Falls back
-    to the first audit record's ts for legacy snapshots without
-    an anchor.
-    """
-    if not audit:
-        return f'<em>{html.escape(_t(locale, "noAudit"))}</em>'
-
-    if base_ts is None:
-        base_ts = next(
-            (r.get("ts") for r in audit
-             if isinstance(r.get("ts"), (int, float))),
-            None,
-        )
-
-    by_set: dict[int, list[dict]] = {}
-    orphans: list[dict] = []
-    for record in audit:
-        set_num = _result_set(record)
-        target = by_set.setdefault(set_num, []) if set_num else orphans
-        target.append(record)
-
-    blocks: list[str] = []
-    ordered_keys = [k for k in range(1, set_count + 1) if k in by_set]
-    # Include any audit-mentioned set numbers above the formal limit
-    # (e.g. data corruption / mode change). They go at the end so the
-    # natural set order still reads top-to-bottom.
-    for k in sorted(by_set.keys()):
-        if k not in ordered_keys:
-            ordered_keys.append(k)
-
-    def _render_li(record: dict) -> str:
-        rel = _format_relative_ts(record.get("ts"), base_ts)
-        label = _action_label(record, locale)
-        running = _running_score_pair(record)
-        running_html = (
-            f' <span class="running">({running[0]}–{running[1]})</span>'
-            if running and _is_score_action(record) else ""
-        )
-        # Per-action-type chip: gives the timeline visual hierarchy
-        # without changing the editorial text. Colour is keyed off
-        # the action kind, not the team — team identity is already
-        # encoded in the label and would clash with the
-        # add_set/timeout / serve / reset / score-edit chip palette
-        # if we tried to layer both. ``_collapse_undos`` already
-        # removed every undo pair upstream, so no row that reaches
-        # this function carries an undone state.
-        action = record.get("action", "")
-        params = record.get("params") or {}
-        team = params.get("team")
-        chip_kind, chip_icon = _chip_classifier(action, team)
-        chip_glyph = (
-            f'<span class="chip-glyph chip-glyph-{chip_kind}" '
-            f'aria-hidden="true">{html.escape(chip_icon)}</span>'
-        )
-        return (
-            f'<li class="timeline-li chip-{chip_kind}">{chip_glyph}'
-            f'<span class="ts">{html.escape(rel)}</span>'
-            f'{html.escape(label)}{running_html}</li>'
-        )
-
-    for set_num in ordered_keys:
-        records = by_set[set_num]
-        items = "".join(_render_li(r) for r in records)
-        blocks.append(
-            f'<section class="timeline-set">'
-            f'<h3>{html.escape(_t(locale, "groupedSetLabel", n=set_num))}</h3>'
-            f'<ol>{items}</ol></section>'
-        )
-
-    if orphans:
-        items = "".join(_render_li(r) for r in orphans)
-        blocks.append(
-            f'<section class="timeline-set">'
-            f'<ol>{items}</ol></section>'
-        )
-
-    # Mini legend so the per-action chip palette is decodable at a
-    # glance. The order is the catalogue's declaration order; each
-    # ``legend_key=None`` entry is skipped (e.g. the generic
-    # ``point`` and the ``other`` fallback don't earn a row of
-    # their own — they overlap semantically with the team-bound
-    # points and an unrenderable action respectively).
-    legend_html_parts: list[str] = []
-    for kind, meta in _CHIP_CATALOGUE.items():
-        legend_key = meta.get("legend_key")
-        if not legend_key:
-            continue
-        label = _t(locale, legend_key)
-        legend_html_parts.append(
-            '<span class="timeline-legend-item">'
-            f'<span class="chip-glyph chip-glyph-{kind}" aria-hidden="true">'
-            f'{html.escape(_chip_glyph(kind))}</span>'
-            f'{html.escape(label)}</span>'
-        )
-    blocks.append(
-        f'<div class="timeline-legend">{"".join(legend_html_parts)}</div>',
-    )
-
-    return "".join(blocks) or f'<em>{html.escape(_t(locale, "noAudit"))}</em>'
-
-
-def _render_logo(customization: dict, team: int) -> str:
-    url = _logo_url(customization, team)
-    if not url:
-        return ""
-    return (
-        f'<img class="logo" src="{html.escape(url)}" '
-        f'alt="" loading="lazy" decoding="async" />'
-    )
 
 
 @match_report_router.get(
@@ -1766,239 +328,6 @@ async def match_report(
 # ---------------------------------------------------------------------------
 # Matches index — browseable list of every archived match for an OID
 # ---------------------------------------------------------------------------
-
-_INDEX_TEMPLATE = """<!doctype html>
-<html lang="en">
-<head>
-<meta charset="utf-8">
-<title>Match history — {oid_label}</title>
-<meta name="viewport" content="width=device-width,initial-scale=1">
-<style>
-  :root {{
-    --fg: #1a1a1a;
-    --muted: #666;
-    --border: #d0d0d0;
-    --hover: #f5f5f5;
-  }}
-  * {{ box-sizing: border-box; }}
-  body {{
-    font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, sans-serif;
-    color: var(--fg);
-    margin: 0;
-    padding: 24px;
-    max-width: 960px;
-    margin-left: auto;
-    margin-right: auto;
-    line-height: 1.5;
-  }}
-  header h1 {{ margin: 0 0 4px; font-size: 24px; }}
-  header .meta {{ color: var(--muted); font-size: 14px; margin-bottom: 16px; }}
-  table {{
-    width: 100%;
-    border-collapse: collapse;
-    margin: 16px 0;
-  }}
-  th, td {{
-    text-align: left;
-    padding: 10px 8px;
-    border-bottom: 1px solid var(--border);
-    font-size: 14px;
-  }}
-  th {{ font-weight: 600; color: var(--muted); }}
-  tbody tr:hover {{ background: var(--hover); }}
-  td.score {{
-    text-align: center;
-    font-weight: 600;
-    font-variant-numeric: tabular-nums;
-  }}
-  td.score .winner {{ color: #2e7d32; }}
-  a.report-link, button.row-delete {{
-    display: inline-block;
-    padding: 4px 10px;
-    border: 1px solid var(--border);
-    border-radius: 4px;
-    text-decoration: none;
-    color: var(--fg);
-    font-size: 13px;
-    background: transparent;
-    cursor: pointer;
-    font-family: inherit;
-  }}
-  a.report-link:hover, button.row-delete:hover {{ background: var(--hover); }}
-  button.row-delete {{ color: #b71c1c; border-color: #ef9a9a; }}
-  button.row-delete:hover {{ background: #fdecea; }}
-  th.checkbox-col, td.checkbox-col {{ width: 28px; text-align: center; padding-left: 0; padding-right: 0; }}
-  .toolbar {{
-    display: flex;
-    align-items: center;
-    gap: 12px;
-    margin-bottom: 8px;
-    font-size: 13px;
-    color: var(--muted);
-  }}
-  .toolbar button {{
-    padding: 6px 14px;
-    border-radius: 4px;
-    border: 1px solid var(--border);
-    background: transparent;
-    color: var(--fg);
-    cursor: pointer;
-    font-family: inherit;
-    font-size: 13px;
-  }}
-  .toolbar button.danger {{
-    color: #b71c1c;
-    border-color: #ef9a9a;
-  }}
-  .toolbar button.danger:hover:not(:disabled) {{ background: #fdecea; }}
-  .toolbar button:disabled {{ opacity: 0.5; cursor: not-allowed; }}
-  .empty {{
-    text-align: center;
-    color: var(--muted);
-    font-style: italic;
-    padding: 32px;
-    border: 1px dashed var(--border);
-    border-radius: 6px;
-  }}
-  .footer {{
-    margin-top: 24px;
-    font-size: 12px;
-    color: var(--muted);
-    border-top: 1px solid var(--border);
-    padding-top: 12px;
-  }}
-</style>
-</head>
-<body>
-<header>
-  <h1>Match history</h1>
-  <div class="meta">OID: <code>{oid_label}</code> &middot; {count} match{plural}</div>
-</header>
-
-{body_html}
-
-<div class="footer">
-  Generated by Volley Overlay Control. Each row links to a print-friendly report.
-</div>
-</body>
-</html>
-"""
-
-
-def _render_match_row(summary: dict, token_qs: str) -> str:
-    """Render one row of the index table."""
-    match_id = summary.get("match_id") or ""
-    t1_sets = summary.get("team_1_sets")
-    t2_sets = summary.get("team_2_sets")
-    winner = summary.get("winning_team")
-    ended = _fmt_ts(summary.get("ended_at"))
-    duration = _fmt_seconds(summary.get("duration_s"))
-
-    def _sets_cell(team: int) -> str:
-        v = (t1_sets if team == 1 else t2_sets)
-        text = "—" if v is None else str(v)
-        if winner == team:
-            return f'<span class="winner">{html.escape(text)}</span>'
-        return html.escape(text)
-
-    safe_id = html.escape(match_id)
-    href = f"/match/{safe_id}/report{token_qs}"
-    return (
-        f'<tr data-match-id="{safe_id}">'
-        f'<td class="checkbox-col">'
-        f'<input type="checkbox" class="match-checkbox" '
-        f'aria-label="Select match {safe_id}" '
-        f'data-match-id="{safe_id}"></td>'
-        f"<td>{html.escape(ended)}</td>"
-        f'<td class="score">{_sets_cell(1)} – {_sets_cell(2)}</td>'
-        f"<td>{html.escape(duration)}</td>"
-        f'<td><a class="report-link" href="{href}">View report</a></td>'
-        f'<td><button type="button" class="row-delete" '
-        f'data-match-id="{safe_id}">Delete</button></td>'
-        "</tr>"
-    )
-
-
-# Static client-side script for the index page. Pure JS — does not go
-# through ``str.format``, so curly braces stay single. Reads the
-# operator's token from the page URL (when present) and forwards it to
-# DELETE /matches/{id} as a Bearer header so the request authenticates
-# the same way the page itself did.
-_INDEX_SCRIPT = """
-<script>
-(function() {
-  const url = new URL(window.location.href);
-  const token = url.searchParams.get('token');
-  const tokenQuery = token ? ('?token=' + encodeURIComponent(token)) : '';
-  const headers = {};
-  if (token) headers['Authorization'] = 'Bearer ' + token;
-
-  const checkboxes = () =>
-    Array.from(document.querySelectorAll('input.match-checkbox'));
-  const selectedIds = () =>
-    checkboxes().filter(cb => cb.checked).map(cb => cb.dataset.matchId);
-
-  function refreshToolbar() {
-    const ids = selectedIds();
-    const btn = document.getElementById('delete-selected');
-    const counter = document.getElementById('selected-count');
-    if (btn) btn.disabled = ids.length === 0;
-    if (counter) counter.textContent = ids.length;
-    const all = document.getElementById('select-all');
-    if (all) {
-      const total = checkboxes().length;
-      all.checked = total > 0 && ids.length === total;
-      all.indeterminate = ids.length > 0 && ids.length < total;
-    }
-  }
-
-  async function deleteOne(id) {
-    const res = await fetch(
-      '/matches/' + encodeURIComponent(id) + tokenQuery,
-      { method: 'DELETE', headers }
-    );
-    return res.ok;
-  }
-
-  async function deleteIds(ids) {
-    if (ids.length === 0) return;
-    const label = ids.length === 1 ? '1 match' : (ids.length + ' matches');
-    if (!confirm('Delete ' + label + '? This cannot be undone.')) return;
-    let failed = 0;
-    for (const id of ids) {
-      try {
-        if (!(await deleteOne(id))) failed++;
-      } catch (e) {
-        failed++;
-      }
-    }
-    if (failed > 0) {
-      alert('Failed to delete ' + failed + ' of ' + ids.length + ' matches.');
-    }
-    window.location.reload();
-  }
-
-  document.addEventListener('change', (e) => {
-    if (e.target.matches('input.match-checkbox')) refreshToolbar();
-    if (e.target.id === 'select-all') {
-      checkboxes().forEach(cb => { cb.checked = e.target.checked; });
-      refreshToolbar();
-    }
-  });
-
-  document.addEventListener('click', (e) => {
-    if (e.target.id === 'delete-selected') {
-      deleteIds(selectedIds());
-    } else if (e.target.matches('button.row-delete')) {
-      deleteIds([e.target.dataset.matchId]);
-    }
-  });
-
-  refreshToolbar();
-})();
-</script>
-"""
-
 
 @match_report_router.get(
     "/matches/index.html",
